pyproject.toml
[project]
name = "rag-faq"
version = "0.1.0"
description = "RAG system for NDPA 2023 legal documents using Google Gemini API and Pinecone"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi==0.115.14",
    "mangum==0.19.0",
    "numpy>=1.26.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.0",
    "pinecone-client==3.2.2",
    "pydantic>=2.7.0",
    "pydantic-core>=2.27.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.20",
    "pyyaml>=6.0.3",
    "redis>=7.2.1",
    "requests>=2.32.5",
    "streamlit>=1.45.0",
    "tiktoken>=0.7.0",
    "uvicorn==0.35.0",
    "google-generativeai>=0.4.0",
    "pdfplumber>=0.11.9",
    "datasets>=4.8.5",
]
